Install grub-editenv and fw_printenv/setenv tools.
So that the bootable partition can be changed afterwards.

Note that we are not preventing the conflicting u-boot-tools from
being reinstalled later, which would destroy the grub-mender-grubenv
tools. The proper fix here would be to rename the grub-mender-grubenv
tools, but this a completely separate issue with too many
ramifications to think about now.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>